in the middle of refactoring the first set of tests
