+++ carta/cpp/Tests/quantileTest.cpp
@@ -1,107 +1,46 @@
 #include "catch.h"
-#include "TestRawView.h"
+#include "quantileTestCommon.h"
 #include "core/Algorithms/percentileAlgorithms.h"
-#include "core/Algorithms/percentileManku99.h"
+// #include "core/Algorithms/percentileManku99.h"
 #include <random>
-
-#include "CartaLib/Slice.h"
-
-/** Dummy view with basic slice support */
-
-class TestRawViewSliceStub
-    : public TestRawView<double>
-{
-public:
-    TestRawViewSliceStub(std::vector<double> data, std::vector<int> dims) : TestRawView(data, dims) {
-    }
-
-    virtual RawViewInterface *
-    getView(const SliceND & sliceInfo) override {
-        // this is hardcoded for this test case only
-        // it implements exactly one kind of slice, which is what is needed by the quantile algorithm
-        // i.e. one frame, and we assume that the image is 3D and the 3rd dimension is the spectral axis
-        // this is also not particularly efficient and should not be used for large test data
-        SliceND::ApplyResult ar = sliceInfo.apply(m_viewDims);
-        
-        size_t frame = ar.dims()[2].start;
-        size_t frameSize = m_viewDims[0] * m_viewDims[1];
-        std::vector<int> frameDims = {m_viewDims[0], m_viewDims[1]};
-        
-        size_t start = frameSize * frame;
-        size_t end = frameSize * (frame + 1);
-        
-        std::vector<double> frameData;
-        
-        for (size_t i = start; i < end; i++) {
-            frameData.push_back(data[i]);
-        }
-        
-        return new TestRawViewSliceStub(frameData, frameDims);
-    }
-};
-
-/** Dummy converters */
-
-class FrameDependentConverter : public Carta::Lib::IntensityUnitConverter {
-public:
-    FrameDependentConverter() 
-        : Carta::Lib::IntensityUnitConverter("DUMMY_FROM", "DUMMY_TO", 2, true, "DUMMY_DEPENDENT_CONVERSION") {
-    }
-
-    double _frameDependentConvert(const double y_val, const double x_val) override {
-        return y_val / pow(x_val, 2);
-    }
-
-    double _frameDependentConvertInverse(const double y_val, const double x_val) override {
-        return y_val * pow(x_val, 2);
-    }
-};
-
-class FrameIndependentConverter : public Carta::Lib::IntensityUnitConverter {
-public:
-    FrameIndependentConverter() 
-        : Carta::Lib::IntensityUnitConverter("DUMMY_FROM", "DUMMY_TO", 2, false, "NONE") {
-    }
-};
-
 
 TEST_CASE( "Quantile algorithm test", "[quantile]" ) {
 
-    std::vector<int> dims = {200, 200, 10};
-    int size = std::accumulate (begin(dims), end(dims), 1, [](int a, int& b){ return b*a; });
-    std::vector<double> data(size);
-    
-    std::iota(data.begin(), data.end(), 1); // we start at 1 because we can't calculate a percentile error if the expected value is 0
-    // shuffle randomly because ordered data breaks the Manku algorithm
-    // but don't seed the generator, to make the shuffle deterministic for the frame-dependent conversion test
-    std::shuffle(data.begin(), data.end(), std::mt19937{});
-    
-    Carta::Lib::NdArray::RawViewInterface * rawView = new TestRawViewSliceStub(data, dims);
-    Carta::Lib::NdArray::Double doubleView(rawView, false);
-        
-    const std::vector<double> percentiles = {0, 0.01, 0.1, 0.9, 0.99, 1};
-
-    std::map <double, double> exactSolution = {
-        {0, 1},
-        {0.01, 4000},
-        {0.1, 40000},
-        {0.9, 360000},
-        {0.99, 396000},
-        {1, 400000}
-    };
-
-    std::map <double, double> exactSolutionFrameDependent = {
-        {0, 0.015625},
-        {0.01, 103.3},
-        {0.1, 1035.14},
-        {0.9, 80052},
-        {0.99, 360171},
-        {1, 399994}
-    };
-        
-    std::vector<double> dummyHzValues(10);
-    std::iota(dummyHzValues.begin(), dummyHzValues.end(), 1); // start with 1 because we divide by these in the dummy converter
-    int spectralIndex = 2;
+//     std::vector<int> dims = {200, 200, 10};
+//     int size = std::accumulate (begin(dims), end(dims), 1, [](int a, int& b){ return b*a; });
+//     std::vector<double> data(size);
+//     
+//     std::iota(data.begin(), data.end(), 1); // we start at 1 because we can't calculate a percentile error if the expected value is 0
+//     // shuffle randomly because ordered data breaks the Manku algorithm
+//     // but don't seed the generator, to make the shuffle deterministic for the frame-dependent conversion test
+//     std::shuffle(data.begin(), data.end(), std::mt19937{});
+//     
+//     Carta::Lib::NdArray::RawViewInterface * rawView = new TestRawViewSliceStub(data, dims);
+//     Carta::Lib::NdArray::Double doubleView(rawView, false);
+//         
+//     const std::vector<double> percentiles = {0, 0.01, 0.1, 0.9, 0.99, 1};
+// 
+//     std::map <double, double> exactSolution = {
+//         {0, 1},
+//         {0.01, 4000},
+//         {0.1, 40000},
+//         {0.9, 360000},
+//         {0.99, 396000},
+//         {1, 400000}
+//     };
+// 
+//     std::map <double, double> exactSolutionFrameDependent = {
+//         {0, 0.015625},
+//         {0.01, 103.3},
+//         {0.1, 1035.14},
+//         {0.9, 80052},
+//         {0.99, 360171},
+//         {1, 399994}
+//     };
+//         
+//     std::vector<double> dummyHzValues(10);
+//     std::iota(dummyHzValues.begin(), dummyHzValues.end(), 1); // start with 1 because we divide by these in the dummy converter
+//     int spectralIndex = 2;
 
     SECTION( "Exact algorithm, no conversion") {
         std::map <double, double> intensities = Carta::Core::Algorithms::percentile2pixels(doubleView, percentiles);
@@ -127,7 +66,7 @@ TEST_CASE( "Quantile algorithm test", "[quantile]" ) {
         // This should remain unchanged; the constant multiplier is not applied within the quantile function
         REQUIRE(intensities == exactSolution);
     }
-
+/*
     SECTION( "Manku 99 algorithm, no conversion") {
         std::map <double, double> intensities =  Carta::Core::Algorithms::percentile2pixels_approximate_manku99(doubleView, percentiles);
         REQUIRE(intensities.size() == percentiles.size());
@@ -167,7 +106,7 @@ TEST_CASE( "Quantile algorithm test", "[quantile]" ) {
             //qDebug() << error << fractionalError;
             REQUIRE(fractionalError < 0.03); // 3%; overly generous, but covers expected spikes
         }
-    }
+    }*/
     
     SECTION("TestRawViewSliceStub implementation") {
         std::vector<double> smallData(2 * 3 * 4);
