+++ carta/cpp/plugins/PercentileManku99/testPercentileManku99.cpp
@@ -1,0 +1,10 @@
+#include "PercentileManku99.h"
+#include <QtTest/QtTest>
+
+class TestPercentileManku99: public QObject
+{
+    Q_OBJECT
+private slots:
+    void test_quantiles_data();
+    void test_quantiles();
+};
